Add support for interruptible handlers
These handlers can hand back control to the server and allow for further
requests to be handled by using `yield`. Typical usecases would be file
uploads and downloads, during which the server would otherwise be blocked.
Handlers without `yield` will continue to run without any change!

Interruptible handlers will work, but have no non-blocking effect in the
default or development servers. Web applications will need to be started
with `-m async`. Because it has not been tested out in the wild yet, this
async server is not the default