Run the PHP development webserver as a backend, forwarding requests to it
To start, this is a simple reverse proxy setup, which is exactly what happens
for HTTP request. Websockets, on the other hand, are kept alive in this server,
and its messages are translated to server-sent events, before being passed to
the backend, which doesn't support this protocol.

This enables the highly effective "save-rerun" developer experience with no
intermediary steps like restarting the server (and potentially losing state),
or compiling (even with the XP Compiler in place, this will happen "just in
time").

We could have chosen any wire format to send the messages back and forth but
chose to go with something well-established and standardized, in this case
opting for https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events